refactor(library): move http class back to separate file (#16)
